refactor(memory): make generation-aware ledger view canonical
